Modify Pulsar class to use SI units and add the capability to return P_dot
Updated the Pulsar class to integrate astropy constants so that all values are calculated and returned in SI units rather than cgs.

Modified the P_dot calculation to use equation from Chattopadhyay et. al. 2020 because the previous equation's units were not working out. 

Added P_dot and alpha to the __init__ function for easier calculation of P_dot and so P_dot can be returned in the history object.

Note: The calculations for the change in angular momentum in RLO_evolve_CMC and RLO_evolve_COMPAS require further review because the units are not matching up between terms.